Enemy cooldown displayed below health

+++ Assets/Scripts/HealthBar.cs
@@ -7,12 +7,12 @@ public class HealthBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetHealth(int health)
+    public void SetHealth(float health)
     {
         slider.value = health;
     }
 
-    public void SetMaxHealth(int max)
+    public void SetMaxHealth(float max)
     {
         slider.maxValue = max;
         slider.value = max;
